services/vanna/requirements-alternative.txt
# Alternative requirements.txt using psycopg (pure Python, no Rust needed)
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
psycopg[binary]==3.2.0
python-dotenv==1.0.1
groq==0.11.0
sqlalchemy==2.0.36

